requirements-worker.txt
requests
python-dotenv
runpod
